e2e-coref/requirements.txt
tensorflow-gpu==1.13.1
tensorflow-hub
h5py
nltk
pyhocon
scipy
sklearn
keras
tqdm
